+++ src/RollTypes.js
@@ -1,3 +1,9 @@
-export const physicalDefense = 'Physical Defense';
-export const weaponAttack = 'Weapon Attack';
-export const weaponDamage = 'Weapon Damage';
+const physicalDefense = 'Physical Defense';
+const weaponAttack = 'Weapon Attack';
+const weaponDamage = 'Weapon Damage';
+
+export default {
+  physicalDefense,
+  weaponAttack,
+  weaponDamage,
+};
