+++ src/action/Stand.js
@@ -1,0 +1,7 @@
+import Action from './Action';
+
+export default class Stand extends Action {
+  constructor(character, characters, nodes) {
+    super('Stand', character, characters, nodes);
+  }
+}
