+++ src/action/index.js
@@ -1,0 +1,17 @@
+import Advance from './Advance';
+import Attack from './Attack';
+import Flee from './Flee';
+import Move from './Move';
+import PowerAttack from './PowerAttack';
+import Rest from './Rest';
+import Stand from './Stand';
+
+export default {
+  Advance,
+  Attack,
+  Flee,
+  Move,
+  PowerAttack,
+  Rest,
+  Stand,
+};
