+++ src/character/npc/index.js
@@ -1,0 +1,7 @@
+import Archer from './Archer';
+import Guard from './Guard';
+
+export default {
+  Archer,
+  Guard,
+};
